+++ priv/templates/brando.install/migrations/20260922000000_brando_155_add_ai_context_fields_to_seo.exs
@@ -1,9 +1,0 @@
-defmodule Brando.Repo.Migrations.AddAiContextFieldsToSEO do
-  use Ecto.Migration
-
-  def change do
-    alter table(:sites_seos) do
-      add :ai_context_fields, :map
-    end
-  end
-end